backend_service/requirements.txt

fastapi
uvicorn
sqlalchemy
pymysql
pydantic
python-multipart
minio
requests
python-dotenv
python-docx
PyPDF2
scikit-learn
